% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecfr.R
\name{ecfr_titles}
\alias{ecfr_titles}
\title{List all CFR titles}
\usage{
ecfr_titles()
}
\value{
tibble: number (integer), name (character),
latest_amended_on (Date), latest_issue_date (Date),
up_to_date_as_of (Date), reserved (logical)
}
\description{
Returns the 50 titles of the Code of Federal Regulations
with their latest amendment dates.
}
