% Generated by build script
\name{ecfr_versions}
\alias{ecfr_versions}
\title{Get change history for a CFR title}
\usage{
ecfr_versions()
}
\arguments{
\item{title_number}{CFR title number (1-50)}
\item{since}{Only show changes since this date (Date or "YYYY-MM-DD")}
}
\value{tibble: date (Date), identifier, name, part, type,}
\description{Get change history for a CFR title}
